tests/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
discord.py>=2.3.0
fastapi>=0.100.0
httpx>=0.24.1
aiohttp>=3.8.5
docker>=6.1.3
pyyaml>=6.0.1
requests>=2.31.0
